refactor: #1424  Add parent class to contribution events
