+++ src/main/java/ai/elimu/model/contributor/ContributionEvent.java
@@ -1,7 +1,7 @@
 package ai.elimu.model.contributor;
 
 import ai.elimu.model.BaseEntity;
-import ai.elimu.model.v2.enums.Platform;
+import ai.elimu.model.enums.Platform;
 import java.util.Calendar;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
