+++ src/main/java/ai/elimu/util/db/DbContentImportHelper.java
@@ -42,7 +42,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.web.context.WebApplicationContext;
 import ai.elimu.dao.LetterSoundCorrespondenceDao;
-import ai.elimu.model.v2.enums.Platform;
+import ai.elimu.model.enums.Platform;
 
 public class DbContentImportHelper {
