+++ src/main/java/ai/elimu/web/content/storybook/StoryBookCreateController.java
@@ -13,6 +13,7 @@ import ai.elimu.model.content.multimedia.Image;
 import ai.elimu.model.contributor.Contributor;
 import ai.elimu.model.contributor.StoryBookContributionEvent;
 import ai.elimu.model.enums.ContentLicense;
+import ai.elimu.model.enums.Platform;
 import ai.elimu.model.v2.enums.ReadingLevel;
 import ai.elimu.util.SlackHelper;
 import ai.elimu.web.context.EnvironmentContextLoaderListener;
