+++ src/main/java/ai/elimu/web/content/storybook/StoryBookCreateFromEPubController.java
@@ -17,7 +17,7 @@ import ai.elimu.model.content.StoryBookParagraph;
 import ai.elimu.model.content.multimedia.Image;
 import ai.elimu.model.contributor.Contributor;
 import ai.elimu.model.contributor.StoryBookContributionEvent;
-import ai.elimu.model.v2.enums.Platform;
+import ai.elimu.model.enums.Platform;
 import ai.elimu.model.v2.enums.content.ImageFormat;
 import ai.elimu.util.ImageColorHelper;
 import ai.elimu.util.ImageHelper;
